add JavaDoc to mutation/selection
