+++ src/test/java/jp/kusumotolab/kgenprog/ga/mutation/selection/EliteAndOldVariantSelectionTest.java
@@ -1,32 +1,31 @@
-package jp.kusumotolab.kgenprog.ga.selection;
+package jp.kusumotolab.kgenprog.ga.mutation.selection;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import org.junit.Before;
 import org.junit.Test;
+import jp.kusumotolab.kgenprog.ga.selection.EliteAndOldVariantSelection;
 import jp.kusumotolab.kgenprog.ga.validation.Fitness;
 import jp.kusumotolab.kgenprog.ga.validation.SimpleFitness;
 import jp.kusumotolab.kgenprog.ga.variant.Variant;
 
-public class DefaultVariantSelectionTest {
+/**
+ * EliteAndOldVariantSelectionに関するテストクラス.
+ */
+public class EliteAndOldVariantSelectionTest {
 
-  private Random random;
-
-  @Before
-  public void setup() {
-    this.random = new Random(0);
-  }
-
+  /**
+   * Variantの算出を正しく行えているかテストする.
+   */
   @Test
   public void testExec() {
     final int variantSize = 5;
-    final DefaultVariantSelection variantSelection = new DefaultVariantSelection(variantSize, random);
+    final EliteAndOldVariantSelection variantSelection = new EliteAndOldVariantSelection(
+        variantSize);
     final List<Variant> variants = new ArrayList<>();
 
     for (int i = 0; i < 10; i++) {
@@ -50,18 +49,28 @@ public class DefaultVariantSelectionTest {
         .containsExactly(0.80d, 0.60d, 0.45d, 0.40d, 0.35d);
   }
 
+  /**
+   * 空のリストに対するVariantの算出を正しく行えているかテストする.
+   * <br>
+   * 結果として空のリストを得られることを期待する.
+   */
   @Test
   public void testExecForEmptyVariants() {
-    final DefaultVariantSelection variantSelection = new DefaultVariantSelection(10, random);
+    final EliteAndOldVariantSelection variantSelection = new EliteAndOldVariantSelection(10);
     final List<Variant> variants1 = Collections.emptyList();
     final List<Variant> variants2 = Collections.emptyList();
     final List<Variant> resultVariants = variantSelection.exec(variants1, variants2);
     assertThat(resultVariants).hasSize(0);
   }
 
+  /**
+   * NaNが含まれるリストに対するVariantの算出を正しく行えているかテストする.
+   * <br>
+   * 結果としてNaNよりも通常の値が優先されることを期待する.
+   */
   @Test
   public void testExecForNan() {
-    final DefaultVariantSelection variantSelection = new DefaultVariantSelection(10, random);
+    final EliteAndOldVariantSelection variantSelection = new EliteAndOldVariantSelection(10);
     final List<Variant> variants = new ArrayList<>();
 
     final List<Variant> nanVariants = IntStream.range(0, 10)
@@ -82,9 +91,14 @@ public class DefaultVariantSelectionTest {
     assertThat(result2.get(0)).isEqualTo(normalVariant);
   }
 
+  /**
+   * NaNが多数含まれるリストに対するVariantの比較を正しく行えているかテストする.
+   * <br>
+   * 結果として個体数を制限し，NaNよりも通常の値が優先されることを期待する.
+   */
   @Test
   public void testExecForNanCompare() {
-    final DefaultVariantSelection variantSelection = new DefaultVariantSelection(10, random);
+    final EliteAndOldVariantSelection variantSelection = new EliteAndOldVariantSelection(10);
 
     final List<Variant> nanVariants = IntStream.range(0, 100)
         .mapToObj(e -> {
@@ -105,14 +119,16 @@ public class DefaultVariantSelectionTest {
   }
 
   /**
-   * 個体の選択を正しく行えているかテストする.<br>
-   * Fitnessが同値の個体があればその中からランダムに選択する.<br>
+   * 個体の選択を正しく行えているかテストする.
+   * <br>
+   * Fitnessが同値の個体同士ではより古い個体を優先して残す.<br>
    * 各Variantにユニークなidを付加し，それを元にして確認を行う.
    */
   @Test
   public void testOrderOfVariants() {
     final int variantSize = 5;
-    final DefaultVariantSelection variantSelection = new DefaultVariantSelection(variantSize, random);
+    final EliteAndOldVariantSelection variantSelection = new EliteAndOldVariantSelection(
+        variantSize);
     final List<Variant> current = new ArrayList<>();
     final List<Variant> generated = new ArrayList<>();
 
@@ -140,14 +156,14 @@ public class DefaultVariantSelectionTest {
 
     assertThat(selectedVariants).hasSize(variantSize)
         .extracting(Variant::getId)
-        .doesNotContainSequence(9L, 19L, 7L, 8L, 17L);
+        .containsSequence(9L, 19L, 7L, 8L, 17L);
   }
 
   /**
-   * Variantを生成するメソッド.Fitnessのみを指定する.
+   * Variantを生成するメソッド. Fitnessのみを指定する.
    *
-   * @param fitness Variantの持つFitness
-   * @return variant 生成したVariant
+   * @param fitness
+   * @return variant
    */
   private Variant createVariant(final Fitness fitness) {
     final Variant variant = new Variant(0, 0, null, null, null, fitness, null, null);
@@ -155,11 +171,11 @@ public class DefaultVariantSelectionTest {
   }
 
   /**
-   * Variantを生成するメソッド.FitnessとIdを指定する.
+   * Variantを生成するメソッド. FitnessとIdを指定する.
    *
-   * @param fitness Variantの持つFitness
-   * @param id Variantに固有の値
-   * @return variant 生成したVariant
+   * @param fitness
+   * @param id
+   * @return variant
    */
   private Variant createVariant(final Fitness fitness, final int id) {
     final Variant variant = new Variant(id, 0, null, null, null, fitness, null, null);
